requirements.txt
tensorflow
matplotlib
numpy
keras
pyyaml
tqdm
opencv-python==4.5.5.62
opencv-contrib-python==4.5.5.62
tensorboard
scipy
Pillow
